refactor: moved custom dispatch
