+++ src/app/App.tsx
@@ -1,6 +1,6 @@
 import {isAppInitSelector} from "@/features/service/selectors/isAppInitSelector";
 import {authThunkCreator} from "@/features/auth/auth-thunks.ts";
-import {useAppDispatch, useAppSelector} from "@/app/hooks.ts";
+import {useAppDispatch, useAppSelector} from "@/common/hooks/useAppDispatch.ts";
 import {useTheme} from "@/features/theme/hooks/useTheme.ts";
 import {Loading} from "@/components/Loading/Loading.tsx";
 import {Layout} from "@/components/Layout/Layout.tsx";
