+++ src/common/hooks/useAppDispatch.ts
@@ -3,8 +3,6 @@ import {ActionsTypes} from "@/app/action-creators-types.ts";
 import {AppRootStateType} from "@/app/store.ts";
 import {ThunkDispatch} from "redux-thunk";
 
-
-//type AppActionsType = ListsActionTypes | TasksActionTypes | AppReducerActionTypes | AuthActionsType;
 export type AppDispatchType = ThunkDispatch<AppRootStateType, any, ActionsTypes>;
 
 export const useAppSelector: TypedUseSelectorHook<AppRootStateType> = useSelector;
