+++ src/components/Toast/Toast.tsx
@@ -1,6 +1,6 @@
 import {appErrorSelector} from "@/features/service/selectors/appErrorSelector";
 import {setAppError} from "@/features/service/service-reducer.ts";
-import {useAppDispatch, useAppSelector} from "@/app/hooks.ts";
+import {useAppDispatch, useAppSelector} from "@/common/hooks/useAppDispatch.ts";
 import {RiErrorWarningFill} from "react-icons/ri";
 import cls from './Toast.module.scss'
 import {useEffect} from "react";
