+++ src/features/profile/components/PostsBlock/Posts/Posts.tsx
@@ -1,7 +1,7 @@
 import cls from "@/features/profile/components/PostsBlock/Posts/Posts.module.scss";
 import {Post} from "@/features/profile/components/PostsBlock/Posts/Post/Post.tsx";
 import {postsSelector} from "@/features/profile/selectors/postsSelector";
-import {useAppSelector} from "@/app/hooks.ts";
+import {useAppSelector} from "@/common/hooks/useAppDispatch.ts";
 
 export const Posts = () => {
