+++ src/pages/Settings/Settings.tsx
@@ -3,7 +3,7 @@ import {switchDependencyOnSystem, switchTheme, ThemeVariantType} from '@/feature
 import {themeSelector} from "@/features/theme/selectors/themeSelector";
 import { PageTemplate } from '@/components/PageTemplate/PageTemplate';
 import {saveThemeToLS} from "@/common/utils/locatStorageUtils.ts";
-import {useAppDispatch, useAppSelector} from "@/app/hooks.ts";
+import {useAppDispatch, useAppSelector} from "@/common/hooks/useAppDispatch.ts";
 import cls from '@/pages/Settings/Settings.module.scss';
 import { RiSettingsLine } from 'react-icons/ri';
 import { RiMoonFill } from 'react-icons/ri';
